BREAKING refactor(io): single-export files
