+++ io/mod.ts
@@ -9,13 +9,20 @@
  * @module
  */
 
-export * from "./buffer.ts";
-export * from "./readers.ts";
-export {
-  copyN,
-  readInt,
-  readLong,
-  readShort,
-  sliceLongToBytes,
-} from "./util.ts";
-export * from "./writers.ts";
+export * from "./buf_reader.ts";
+export * from "./buf_writer.ts";
+/** @deprecated (will be removed after 0.170.0) `export * from "./buffer.ts";` upon removal */
+export { Buffer } from "./buffer.ts";
+export * from "./copy_n.ts";
+export * from "./limited_reader.ts";
+export * from "./multi_reader.ts";
+export * from "./read_delim.ts";
+export * from "./read_int.ts";
+export * from "./read_lines.ts";
+export * from "./read_long.ts";
+export * from "./read_range.ts";
+export * from "./read_short.ts";
+export * from "./read_string_delim.ts";
+export * from "./slice_long_to_bytes.ts";
+export * from "./string_reader.ts";
+export * from "./string_writer.ts";
