+++ io/string_writer_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { assertEquals } from "../testing/asserts.ts";
-import { StringWriter } from "./writers.ts";
-import { StringReader } from "./readers.ts";
-import { copyN } from "./util.ts";
+import { StringWriter } from "./string_writer.ts";
+import { StringReader } from "./string_reader.ts";
+import { copyN } from "./copy_n.ts";
 import { copy } from "../streams/copy.ts";
 
 Deno.test("ioStringWriter", async function () {
