--- io/test.ts
@@ -1,2 +1,0 @@
-// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-import "./mod.ts";
